Moved more doc fncs from CfgProj to DocProj
